# Limit the number of symbol completion candidates and make it faster