The number of symbol candidates can be huge.

Reduce calling `symbol.inspect` to make symbol search faster
```ruby
symbols = 100000.times.map{rand(36**rand(5..25)).to_s(36).to_sym};
symbols.map{_1.inspect[1..]}.select{_1.start_with? 'ab'};
# processing time: 0.063232s
symbols.select{_1.start_with? 'ab'}.filter_map{a=_1.inspect[1..];a if a.start_with? 'ab'};
# processing time: 0.017167s
```

Limit the number of symbol completion candidates. Avoid calling `large_symbol_array.sort` and use `min(n)+max(n).reverse`. This will also improve performance of rendering completion dialog in Reline.
```ruby
symbols.size
# => 100000
symbols.sort;
# processing time: 0.061497s
symbols.min(50)+symbols.max(50).reverse;
# processing time: 0.013646s
```